feat: APIをランキング画面につなげました

+++ app/src/main/java/com/example/drivequest/data/remote/api/drivehistory/model/DriveHistoryResponse.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 data class DriveHistoryResponse (
     val drivingHistoryId: String,
     val userId: String,
-    val drivingMileage: Int,
+    val drivingMileage: Double,
     val durationTime: Int,
     val routeData: String?,
     val createdAt: String,
